# docs/swagger.yaml
consumes:
- application/json
definitions:
  ginx.Result:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  web.SignUpReq:
    properties:
      mobile:
        type: string
      nick_name:
        type: string
      signup_token:
        type: string
    type: object
  web.editReq:
    properties:
      city:
        type: string
      constellation:
        type: integer
      gender:
        type: string
      name:
        type: string
      province:
        type: string
      signature:
        type: string
    type: object
  web.loginReq:
    properties:
      code:
        type: string
    type: object
  web.reqUploadToken:
    properties:
      file_ext:
        type: string
      file_type:
        type: string
    type: object
info:
  contact: {}
  title: 你好同城后端API
paths:
  /files/upload_token:
    post:
      consumes:
      - application/json
      description: 文件上传token获取接口
      parameters:
      - description: 上传文件token获取参数
        in: body
        name: UploadToken
        required: true
        schema:
          $ref: '#/definitions/web.reqUploadToken'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":xxx,"data":{},"msg":"xxx"}'
          schema:
            $ref: '#/definitions/ginx.Result'
      summary: 文件上传token获取接口
      tags:
      - 文件相关接口
  /users/login:
    post:
      consumes:
      - application/json
      description: 登录成功返回的token放在响应的header的x-jwt-token里面，登录之后的后续访问需要带上token，放在请求的header里面的Authorization。
      parameters:
      - description: 微信登录的临时登录凭证
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/web.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":xxx,"data":{},"msg":"xxx"}'
          schema:
            $ref: '#/definitions/ginx.Result'
      summary: 用户登录接口
      tags:
      - 用户相关接口
  /users/profile:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户信息更新参数
        in: body
        name: edit
        required: true
        schema:
          $ref: '#/definitions/web.editReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":xxx,"data":{},"msg":"xxx"}'
          schema:
            $ref: '#/definitions/ginx.Result'
      summary: 用户信息更新接口
      tags:
      - 用户相关接口
  /users/signup:
    post:
      consumes:
      - application/json
      description: 用户注册接口
      parameters:
      - description: 注册参数
        in: body
        name: signup
        required: true
        schema:
          $ref: '#/definitions/web.SignUpReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":xxx,"data":{},"msg":"xxx"}'
          schema:
            $ref: '#/definitions/ginx.Result'
      summary: 用户注册接口
      tags:
      - 用户相关接口
produces:
- application/json
schemes:
- http
- https
swagger: "2.0"